Add "RailsStyle" connection object with overrides
Overrides:
- prepackage_and_send
- document_details

RightSignature "thinks" in terms of XML-style data structures (this
mostly means arrays of objects, instead of hashes). This can be awkward
to integrate, specifically with Rails. This begins to allow Rails to
think in classic Ruby data structures, and handle the translation to
the XML style for you, behind the scenes.